Use the configured ruby version instead of `current` for `WhenDecomposer`
This one has been bothering me for a while. When requiring `parser/current`, it checks
if the executing ruby patch version is the expected one. If it isn't, it will show a warning:
> warning: parser/current is loading parser/ruby33, which recognizes 3.3.4-compliant syntax, but you are running 3.3.3

RuboCop doesn't show this because it requires specific parser versions that can't possibly warn about this.

This is also technically more correct, though I'm not quite sure if that even matters